# Conflicts:
#	src/lib/app/eventTypes.ts
#	src/lib/server/random.ts
#	src/lib/ui/ui.ts
#	src/lib/vocab/community/communityEvents.ts
#	src/lib/vocab/community/communityRepo.ts
#	src/lib/vocab/community/communityServices.ts
#	src/lib/vocab/membership/membershipRepo.ts
#	src/lib/vocab/membership/membershipServices.ts

+++ src/lib/vocab/community/communityServices.ts
@@ -20,7 +20,7 @@ import {
 	ReadCommunity,
 	UpdateCommunitySettings,
 	DeleteCommunity,
-} from '$lib/vocab/community/community.events';
+} from '$lib/vocab/community/communityEvents';
 import {toDefaultCommunitySettings} from '$lib/vocab/community/community.schema';
 
 const log = new Logger(gray('[') + blue('communityServices') + gray(']'));
@@ -154,6 +154,21 @@ export const deleteCommunityService: Service<DeleteCommunityParams, DeleteCommun
 	{
 		event: DeleteCommunity,
 		perform: async ({repos, params}) => {
+			const communityResult = await repos.community.findById(params.community_id);
+			if (!communityResult.ok) {
+				return {
+					ok: false,
+					status: 404,
+					message: communityResult.message || 'issue finding community',
+				};
+			}
+			if (communityResult.value.type === 'personal') {
+				return {
+					ok: false,
+					status: 405,
+					message: 'cannot delete personal community',
+				};
+			}
 			const deleteResult = await repos.community.deleteById(params.community_id);
 			if (!deleteResult.ok) {
 				return {ok: false, status: 500, message: deleteResult.message || 'unknown error'};
